Moved ThreadLocalSingleton to separate file.
