+++ that_depends/providers/__init__.py
@@ -8,10 +8,11 @@ from that_depends.providers.context_resources import (
     container_context,
 )
 from that_depends.providers.factories import AsyncFactory, Factory
+from that_depends.providers.local_singleton import ThreadLocalSingleton
 from that_depends.providers.object import Object
 from that_depends.providers.resources import Resource
 from that_depends.providers.selector import Selector
-from that_depends.providers.singleton import AsyncSingleton, Singleton, ThreadLocalSingleton
+from that_depends.providers.singleton import AsyncSingleton, Singleton
 
 
 __all__ = [
